refactor: Перенёс ParticipationRequestDto
