+++ ewm-main-service/src/main/java/ru/practicum/dto/EventRequestStatusUpdateResult.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.practicum.participation.dto.ParticipationRequestDto;
 
 import java.util.List;
